story-4 add podcast scope
