--- Silkrode/app/src/main/java/personal/ivan/silkrode/navigation/podcast/view/ChannelListFragment.kt
@@ -1,19 +1,0 @@
-package personal.ivan.silkrode.navigation.podcast.view
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import personal.ivan.silkrode.R
-
-class ChannelListFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_channel_list, container, false)
-    }
-}
