+++ Silkrode/app/src/main/res/navigation/nav_podcast.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/channelListFragment"
-        android:name="personal.ivan.silkrode.navigation.podcast.view.ChannelListFragment"
+        android:name="personal.ivan.silkrode.navigation.podcast.view.fragment.ChannelListFragment"
         android:label="ChannelListFragment"
         tools:layout="@layout/fragment_channel_list">
         <action
@@ -20,7 +20,7 @@
     </fragment>
     <fragment
         android:id="@+id/programListFragment"
-        android:name="personal.ivan.silkrode.navigation.podcast.view.ProgramListFragment"
+        android:name="personal.ivan.silkrode.navigation.podcast.view.fragment.ProgramListFragment"
         android:label="ProgramListFragment"
         tools:layout="@layout/fragment_program_list">
         <action
@@ -33,7 +33,7 @@
     </fragment>
     <fragment
         android:id="@+id/playFragment"
-        android:name="personal.ivan.silkrode.navigation.podcast.view.PlayFragment"
+        android:name="personal.ivan.silkrode.navigation.podcast.view.fragment.PlayFragment"
         android:label="PlayFragment"
         tools:layout="@layout/fragment_play" />
 </navigation>
